use new scanner lib

+++ mrsystem/src/basic_libs/scanner/scan_eol.c
@@ -10,8 +10,9 @@ static char NextChar(Scanner *Handle)
 {
    if (Handle->AktZeichen>=Handle->AnzZeichen)
    {
-      Handle->AnzZeichen = fread(Handle->Puffer, 1, SCAN_BLOCK_LAENGE,
-                                 Handle->ScanFile);
+      Handle->AnzZeichen = (int)fread(Handle->Puffer, 1,
+                                      SCAN_BLOCK_LAENGE,
+                                      Handle->ScanFile);
       Handle->AktZeichen = 0;
    }
    if (Handle->AktZeichen < Handle->AnzZeichen)
@@ -21,8 +22,9 @@ static char NextChar(Scanner *Handle)
       {
          if (Handle->AktZeichen > Handle->AnzZeichen)
          {
-            Handle->AnzZeichen = fread(Handle->Puffer, 1, SCAN_BLOCK_LAENGE,
-                                       Handle->ScanFile);
+            Handle->AnzZeichen = (int)fread(Handle->Puffer, 1,
+                                            SCAN_BLOCK_LAENGE,
+                                            Handle->ScanFile);
             Handle->AktZeichen = 0;
          }
          if ((Handle->AnzZeichen > 0) &&
